refactor out generate columns func to separate utils file
